Split package invocation service
