+++ packages/worker/src/package-invocations/service.ts
@@ -1,1299 +1,50 @@
-import { type ContentBlock } from '@modelcontextprotocol/sdk/types.js'
-import { normalizePackageInvocationExportName } from '@kody-internal/shared/public-urls.ts'
-import { toHex } from '@kody-internal/shared/hex.ts'
-import { type RemoteConnectorRef } from '@kody-internal/shared/remote-connectors.ts'
-import { getExecutionErrorDetails } from '#mcp/executor.ts'
-import { persistableExecutionArtifacts } from '#mcp/downstream-mcp-result.ts'
-import { createMcpCallerContext } from '#mcp/context.ts'
-import { canonicalJsonStringify } from '@kody-internal/shared/canonical-json.ts'
-import { toJsonSafeValue } from '@kody-internal/shared/json-safe-value.ts'
-import { getErrorMessage } from '@kody-internal/shared/error-message.ts'
+import { type createMcpCallerContext } from '#mcp/context.ts'
 import {
-	runBundledModuleWithRegistry,
-	type PackageEventDispatchInput,
 	type PackageEventTools,
-	type PackageInvokeCheckResult,
-	type PackageInvokeContract,
-	type PackageInvokeInput,
-	type PackageInvokeNormalizedInput,
 	type PackageInvokeTools,
 } from '#mcp/run-kody-registry.ts'
-import {
-	type PackageRuntimeDebugContext,
-	type PackageRuntimeSurface,
-} from '#worker/package-runtime/package-runtime-debug.ts'
-import {
-	listSavedPackagesByUserId,
-	getSavedPackageById,
-	getSavedPackageByKodyId,
-} from '#worker/package-registry/repo.ts'
-import {
-	loadPackageManifestBySourceId,
-	loadPackageSourceBySourceId,
-} from '#worker/package-registry/source.ts'
+import { type PackageRuntimeDebugContext } from '#worker/package-runtime/package-runtime-debug.ts'
 import { type SavedPackageRecord } from '#worker/package-registry/types.ts'
 import {
-	buildPackageSearchProjection,
-	listPackageEmittedEvents,
-	listPackageSubscriptions,
-	normalizePackageWorkspacePath,
-	resolvePackageExportPath,
-	type PackageExportProjection,
-} from '#worker/package-registry/manifest.ts'
-import { typecheckPackageEntrypointsFromSourceFiles } from '#worker/repo/checks.ts'
+	normalizeExportName,
+	packageInvocationScopeWildcard,
+	type PackageInvocationRequest,
+	type PackageInvocationResponse,
+	type PackageInvocationTokenScope,
+	type PackageRuntimeContext,
+	type PackageRuntimeToolFactories,
+} from './common.ts'
+import { invokePackageExportWithToolFactories } from './http-invoke.ts'
 import {
-	loadPublishedBundleArtifactByIdentity,
-	persistPublishedBundleArtifact,
-} from '#worker/package-runtime/published-bundle-artifacts.ts'
-import { packageWorkflowInvocationSource } from '#worker/package-runtime/package-invocation-sources.ts'
-import { assertPublishedSourceCanRebuildWithoutInstallingDeps } from '#worker/package-runtime/published-source-dependencies.ts'
-import { recordAgentPackageConversationUse } from '#worker/usage/agent-package-conversation-uses.ts'
+	createExecutePackageInvokeToolsWithToolFactories,
+	createPackageRuntimeInvokeToolsWithToolFactories,
+} from './runtime-tool-factories.ts'
 import {
-	buildPackageSubscriptionArtifactName,
-	normalizePackageSubscriptionTopic,
-} from '#worker/package-runtime/subscription-artifacts.ts'
-import { getEntitySourceById } from '#worker/repo/entity-sources.ts'
-import { type EntitySourceRow } from '#worker/repo/types.ts'
-import {
-	getEmailAttachmentById,
-	getEmailMessageById,
-	getEmailMessageWithAttachmentsById,
-} from '#worker/email/repo.ts'
-import { buildPackageStorageId } from '#worker/storage-runner.ts'
-import {
-	getPackageInvocationByKey,
-	insertPackageInvocationRow,
-	updatePackageInvocationResult,
-	type PackageInvocationStoredResponse,
-} from './repo.ts'
+	createPackageEventToolsWithToolFactories,
+	invokePackageSubscriptionWithToolFactories,
+} from './subscription-dispatch.ts'
 
-export type PackageInvocationTokenScope = {
-	tokenId: string
-	userId: string
-	email: string
-	displayName: string
-	packageIds?: Array<string>
-	packageKodyIds?: Array<string>
-	exportNames?: Array<string>
-	sources?: Array<string>
-	remoteConnectors?: Array<RemoteConnectorRef>
+export {
+	normalizeExportName,
+	packageInvocationScopeWildcard,
+	type PackageInvocationRequest,
+	type PackageInvocationResponse,
+	type PackageInvocationTokenScope,
 }
 
-export type PackageInvocationRequest = {
-	packageIdOrKodyId: string
-	exportName: string
-	params?: Record<string, unknown>
-	idempotencyKey: string
-	source?: string | null
-	topic?: string | null
-}
-
-export type PackageInvocationResponse = PackageInvocationStoredResponse
-
-type PackageInvocationActor = {
-	tokenId: string
-	userId: string
-	email: string
-	displayName: string
-	remoteConnectors?: Array<RemoteConnectorRef> | null
-}
-
-type PackageModuleSelector =
-	| {
-			kind: 'export'
-			exportName: string
-	  }
-	| {
-			kind: 'subscription'
-			topic: string
-	  }
-
-type PackageModuleResolution = {
-	artifactName: string
-	entryPoint: string
-}
-
-type PackageRuntimeContext = {
-	packageId: string
-	kodyId: string
-	sourceId?: string | null
-}
-
-type LoadedPackageEventSubscription = {
-	savedPackage: SavedPackageRecord
-	subscription: ReturnType<typeof listPackageSubscriptions>[number]
-}
-
-const internalEmailSubscriptionTokenId = 'internal:email-subscriptions'
-const internalPackageEventSubscriptionTokenId = 'internal:package-events'
-const internalPackageRuntimeInvokeTokenId = 'internal:package-runtime'
-const internalExecuteRuntimeInvokeTokenId = 'internal:execute-runtime'
-const maxPackageRuntimeInvokeDepth = 8
-export const packageInvocationScopeWildcard = '*'
-
-export function normalizeExportName(exportName: string) {
-	return normalizePackageInvocationExportName(exportName)
-}
-
-function normalizeNullableString(value: string | null | undefined) {
-	const trimmed = value?.trim()
-	return trimmed && trimmed.length > 0 ? trimmed : null
-}
-
-function readOptionalString(
-	input: PackageInvokeInput,
-	fieldNames: Array<string>,
-) {
-	for (const fieldName of fieldNames) {
-		const value = input[fieldName]
-		if (typeof value === 'string' && value.trim()) return value.trim()
-	}
-	return null
-}
-
-function readSinglePackageIdentifier(
-	input: PackageInvokeInput,
-	operationName = 'packages.invoke',
-) {
-	const candidates = [
-		{
-			kind: 'kodyId' as const,
-			value: readOptionalString(input, ['kodyId', 'kody_id']),
-		},
-		{
-			kind: 'packageId' as const,
-			value: readOptionalString(input, ['packageId', 'package_id']),
-		},
-	].filter(
-		(candidate): candidate is { kind: 'kodyId' | 'packageId'; value: string } =>
-			candidate.value !== null,
-	)
-	const unique = Array.from(
-		new Set(candidates.map((candidate) => candidate.value)),
-	)
-	if (unique.length > 1) {
-		throw new Error(
-			`${operationName} accepts one package identifier. Use kodyId unless you need the saved package id.`,
-		)
-	}
-	const [identifier] = candidates
-	if (!identifier) {
-		throw new Error(`${operationName} requires kodyId or packageId.`)
-	}
-	return identifier
-}
-
-function readPackageInvokeParams(
-	input: PackageInvokeInput,
-	operationName = 'packages.invoke',
-) {
-	const params = input['params']
-	if (params === undefined || params === null) return undefined
-	if (!params || typeof params !== 'object' || Array.isArray(params)) {
-		throw new Error(
-			`${operationName} params must be a JSON object when provided.`,
-		)
-	}
-	return params as Record<string, unknown>
-}
-
-function parsePackageInvokeInput(
-	input: PackageInvokeInput,
-	operationName = 'packages.invoke',
-) {
-	const packageIdentifier = readSinglePackageIdentifier(input, operationName)
-	const exportName = readOptionalString(input, ['exportName', 'export_name'])
-	if (!exportName) {
-		throw new Error(`${operationName} requires exportName.`)
-	}
-	return {
-		packageIdentifier,
-		packageIdOrKodyId: packageIdentifier.value,
-		exportName,
-		params: readPackageInvokeParams(input, operationName),
-		idempotencyKey: readOptionalString(input, [
-			'idempotencyKey',
-			'idempotency_key',
-		]),
-		topic: readOptionalString(input, ['topic']),
-	}
-}
-
-function buildPackageInvocationStorageId(packageId: string) {
-	// Shared with packageStorage() so all surfaces name the same bucket. Since
-	// the ambient `storage` binding was removed from invocation runs, this only
-	// feeds `callerContext.storageContext` (secret scoping and runtime-debug
-	// metadata) — package code reaches the bucket via `packageStorage()`.
-	return buildPackageStorageId(packageId)
-}
-
-function createRepoContext(source: EntitySourceRow) {
-	return {
-		sourceId: source.id,
-		repoId: source.repo_id,
-		sessionId: null,
-		baseCommit: source.published_commit,
-		manifestPath: source.manifest_path,
-		sourceRoot: source.source_root,
-		publishedCommit: source.published_commit,
-		entityKind: source.entity_kind,
-		entityId: source.entity_id,
-	}
-}
-
-function markStoredResponseAsReplayed(
-	response: PackageInvocationStoredResponse,
-) {
-	const body = structuredClone(response.body)
-	const record = body as Record<string, unknown>
-	const existingIdempotency = record['idempotency']
-	if (existingIdempotency && typeof existingIdempotency === 'object') {
-		record['idempotency'] = {
-			...(existingIdempotency as Record<string, unknown>),
-			replayed: true,
-		}
-	} else {
-		record['idempotency'] = { replayed: true }
-	}
-	return {
-		status: response.status,
-		body,
-	} satisfies PackageInvocationStoredResponse
-}
-
-async function createRequestHash(input: {
-	packageId: string
-	exportName: string
-	params?: Record<string, unknown>
-	source: string | null
-	topic: string | null
-}) {
-	const canonical = canonicalJsonStringify(input)
-	const digest = await crypto.subtle.digest(
-		'SHA-256',
-		new TextEncoder().encode(canonical),
-	)
-	return toHex(new Uint8Array(digest))
-}
-
-async function createAutoPackageInvokeIdempotencyKey(input: {
-	callerPackageContext: PackageRuntimeContext | null
-	parentRuntimeDebug: PackageRuntimeDebugContext | null
-	sequence: number
-	request: ReturnType<typeof parsePackageInvokeInput>
-}) {
-	if (!input.callerPackageContext) {
-		return `pkginvoke:${crypto.randomUUID()}`
-	}
-	const parentKey = input.parentRuntimeDebug?.idempotencyKey?.trim()
-	if (!parentKey) {
-		return `pkginvoke:${crypto.randomUUID()}`
-	}
-	const digest = await crypto.subtle.digest(
-		'SHA-256',
-		new TextEncoder().encode(
-			canonicalJsonStringify({
-				callerPackageId: input.callerPackageContext.packageId,
-				parentKey,
-				parentSurface: input.parentRuntimeDebug?.surface ?? null,
-				parentName: input.parentRuntimeDebug?.name ?? null,
-				sequence: input.sequence,
-				packageIdOrKodyId: input.request.packageIdOrKodyId,
-				exportName: normalizeExportName(input.request.exportName),
-				params: input.request.params,
-				topic: input.request.topic,
-			}),
-		),
-	)
-	return [
-		'pkginvoke',
-		input.callerPackageContext.packageId,
-		parentKey,
-		String(input.sequence),
-		toHex(new Uint8Array(digest)).slice(0, 24),
-	].join(':')
-}
-
-function createPackageInvokeCheckFailure(input: {
-	message: string
-	problems: Array<string>
-	contract?: Partial<PackageInvokeContract>
-}): PackageInvokeCheckResult {
-	return {
-		ok: false,
-		message: input.message,
-		problems: input.problems,
-		...(input.contract ? { contract: input.contract } : {}),
-	}
-}
-
-function buildNormalizedPackageInvokeInput(input: {
-	request: ReturnType<typeof parsePackageInvokeInput>
-	exportName: string
-}): PackageInvokeNormalizedInput {
-	return {
-		...(input.request.packageIdentifier.kind === 'kodyId'
-			? { kodyId: input.request.packageIdentifier.value }
-			: { packageId: input.request.packageIdentifier.value }),
-		exportName: input.exportName,
-		...(input.request.params === undefined
-			? {}
-			: { params: input.request.params }),
-		...(input.request.idempotencyKey
-			? { idempotencyKey: input.request.idempotencyKey }
-			: {}),
-		...(input.request.topic ? { topic: input.request.topic } : {}),
-	}
-}
-
-function findPackageExportProjection(input: {
-	exports: Array<PackageExportProjection>
-	exportName: string
-}) {
-	return (
-		input.exports.find(
-			(exportDetail) => exportDetail.subpath === input.exportName,
-		) ?? null
-	)
-}
-
-function buildPackageInvokeCheckWarnings(input: {
-	exportDetail: PackageExportProjection | null
-	sourceLoadFailed: boolean
-}) {
-	const warnings = [
-		'No machine-readable params schema is published for package exports; params were only validated as a JSON object.',
-	]
-	if (input.sourceLoadFailed) {
-		warnings.push(
-			'Source files could not be loaded for metadata extraction; description and type information may be incomplete.',
-		)
-	}
-	if (!input.exportDetail?.typeDefinition) {
-		warnings.push(
-			'No function type definition was found for this export; callable shape could not be statically confirmed.',
-		)
-	}
-	return warnings
-}
-
-async function checkPackageInvokeForRuntime(input: {
-	env: Env
-	baseUrl: string
-	operationName: 'packages.check' | 'packages.invokeChecked'
-	userId: string
-	rawInput: PackageInvokeInput
-}): Promise<PackageInvokeCheckResult> {
-	let request: ReturnType<typeof parsePackageInvokeInput>
-	try {
-		request = parsePackageInvokeInput(input.rawInput, input.operationName)
-	} catch (error) {
-		const message = getErrorMessage(error)
-		return createPackageInvokeCheckFailure({
-			message,
-			problems: [message],
+const packageRuntimeToolFactories: PackageRuntimeToolFactories = {
+	createPackageRuntimeInvokeTools(input) {
+		return createPackageRuntimeInvokeToolsWithToolFactories({
+			...input,
+			toolFactories: packageRuntimeToolFactories,
 		})
-	}
-	const exportName = normalizeExportName(request.exportName)
-	const invoke = buildNormalizedPackageInvokeInput({ request, exportName })
-	const savedPackage = await resolveSavedPackage({
-		db: input.env.APP_DB,
-		userId: input.userId,
-		packageIdOrKodyId: request.packageIdOrKodyId,
-	})
-	if (!savedPackage) {
-		const message = `Saved package "${request.packageIdOrKodyId}" was not found for this user.`
-		return createPackageInvokeCheckFailure({
-			message,
-			problems: [message],
-			contract: { exportName },
+	},
+	createPackageEventTools(input) {
+		return createPackageEventToolsWithToolFactories({
+			...input,
+			toolFactories: packageRuntimeToolFactories,
 		})
-	}
-	const packageContract = {
-		packageId: savedPackage.id,
-		kodyId: savedPackage.kodyId,
-		name: savedPackage.name,
-		sourceId: savedPackage.sourceId,
-		exportName,
-	}
-	let manifestResult: Awaited<ReturnType<typeof loadPackageManifestBySourceId>>
-	try {
-		manifestResult = await loadPackageManifestBySourceId({
-			env: input.env,
-			baseUrl: input.baseUrl,
-			userId: input.userId,
-			sourceId: savedPackage.sourceId,
-		})
-	} catch (error) {
-		const problem = `Could not load current package manifest: ${getErrorMessage(error)}`
-		return createPackageInvokeCheckFailure({
-			message: problem,
-			problems: [problem],
-			contract: packageContract,
-		})
-	}
-	let resolution: PackageModuleResolution
-	try {
-		resolution = resolvePackageModuleResolution({
-			manifest: manifestResult.manifest,
-			selector: {
-				kind: 'export',
-				exportName,
-			},
-		})
-	} catch (error) {
-		const problem = getErrorMessage(error)
-		return createPackageInvokeCheckFailure({
-			message: problem,
-			problems: [problem],
-			contract: {
-				...packageContract,
-				publishedCommit: manifestResult.source.published_commit ?? null,
-			},
-		})
-	}
-	try {
-		await ensureModuleArtifact({
-			env: input.env,
-			baseUrl: input.baseUrl,
-			packageManifest: manifestResult,
-			resolution,
-			savedPackage,
-			selector: {
-				kind: 'export',
-				exportName,
-			},
-			userId: input.userId,
-		})
-	} catch (error) {
-		const problem = `Export "${exportName}" could not be prepared for invocation: ${getErrorMessage(error)}`
-		return createPackageInvokeCheckFailure({
-			message: problem,
-			problems: [problem],
-			contract: {
-				...packageContract,
-				publishedCommit: manifestResult.source.published_commit ?? null,
-				runtimeTarget: resolution.entryPoint,
-			},
-		})
-	}
-	let files: Record<string, string> | undefined
-	let sourceLoadFailed = false
-	try {
-		files = (
-			await loadPackageSourceBySourceId({
-				env: input.env,
-				baseUrl: input.baseUrl,
-				userId: input.userId,
-				sourceId: savedPackage.sourceId,
-			})
-		).files
-	} catch {
-		sourceLoadFailed = true
-	}
-	const projection = buildPackageSearchProjection(
-		manifestResult.manifest,
-		files,
-	)
-	const exportDetail = findPackageExportProjection({
-		exports: projection.exports,
-		exportName,
-	})
-	const warnings = buildPackageInvokeCheckWarnings({
-		exportDetail,
-		sourceLoadFailed,
-	})
-	return {
-		ok: true,
-		invoke,
-		contract: {
-			...packageContract,
-			publishedCommit: manifestResult.source.published_commit ?? null,
-			runtimeTarget: exportDetail?.runtimeTarget ?? resolution.entryPoint,
-			description: exportDetail?.description ?? null,
-			typeDefinition: exportDetail?.typeDefinition ?? null,
-			warnings,
-		},
-	}
-}
-
-async function resolveSavedPackage(input: {
-	db: D1Database
-	userId: string
-	packageIdOrKodyId: string
-}): Promise<SavedPackageRecord | null> {
-	return (
-		(await getSavedPackageById(input.db, {
-			userId: input.userId,
-			packageId: input.packageIdOrKodyId,
-		})) ??
-		(await getSavedPackageByKodyId(input.db, {
-			userId: input.userId,
-			kodyId: input.packageIdOrKodyId,
-		}))
-	)
-}
-
-function scopeAllowsValue(input: { values?: Array<string>; value: string }) {
-	return (
-		input.values?.includes(packageInvocationScopeWildcard) ||
-		input.values?.includes(input.value) ||
-		false
-	)
-}
-
-function tokenAllowsPackage(input: {
-	token: PackageInvocationTokenScope
-	savedPackage: SavedPackageRecord
-}) {
-	return (
-		scopeAllowsValue({
-			values: input.token.packageIds,
-			value: input.savedPackage.id,
-		}) ||
-		scopeAllowsValue({
-			values: input.token.packageKodyIds,
-			value: input.savedPackage.kodyId,
-		})
-	)
-}
-
-function tokenAllowsExport(input: {
-	token: PackageInvocationTokenScope
-	exportName: string
-}) {
-	const exportNames = input.token.exportNames ?? []
-	return exportNames.some(
-		(entry) =>
-			entry === packageInvocationScopeWildcard ||
-			normalizeExportName(entry) === input.exportName,
-	)
-}
-
-function tokenAllowsSource(input: {
-	token: PackageInvocationTokenScope
-	source: string | null
-}) {
-	if (!input.source) return true
-	const sources = input.token.sources ?? []
-	return sources.includes(input.source)
-}
-
-async function ensureModuleArtifact(input: {
-	env: Env
-	baseUrl: string
-	packageManifest?: Awaited<ReturnType<typeof loadPackageManifestBySourceId>>
-	resolution?: PackageModuleResolution
-	savedPackage: SavedPackageRecord
-	selector: PackageModuleSelector
-	userId: string
-}) {
-	const packageManifest =
-		input.packageManifest ??
-		(await loadPackageManifestBySourceId({
-			env: input.env,
-			baseUrl: input.baseUrl,
-			userId: input.userId,
-			sourceId: input.savedPackage.sourceId,
-		}))
-	const resolution =
-		input.resolution ??
-		resolvePackageModuleResolution({
-			manifest: packageManifest.manifest,
-			selector: input.selector,
-		})
-	const loaded = await loadPublishedBundleArtifactByIdentity({
-		env: input.env,
-		userId: input.userId,
-		sourceId: input.savedPackage.sourceId,
-		kind: 'module',
-		artifactName: resolution.artifactName,
-		entryPoint: resolution.entryPoint,
-	})
-	if (loaded?.artifact) {
-		return {
-			artifact: loaded.artifact,
-			source: packageManifest.source,
-			entryPoint: resolution.entryPoint,
-		}
-	}
-	const packageSource = await loadPackageSourceBySourceId({
-		env: input.env,
-		baseUrl: input.baseUrl,
-		userId: input.userId,
-		sourceId: input.savedPackage.sourceId,
-	})
-	const typecheckResult = await typecheckPackageEntrypointsFromSourceFiles({
-		sourceFiles: packageSource.files,
-		entryPoints: [{ path: resolution.entryPoint, includeStorage: true }],
-		emittedEventTopics: Object.keys(packageSource.manifest.kody.emits ?? {}),
-	})
-	if (!typecheckResult.ok) {
-		throw new Error(typecheckResult.message)
-	}
-	assertPublishedSourceCanRebuildWithoutInstallingDeps({
-		sourceFiles: packageSource.files,
-		bundleLabel: `Saved package export "${resolution.artifactName}"`,
-	})
-	const { buildKodyModuleBundle } =
-		await import('#worker/package-runtime/module-graph.ts')
-	const bundle = await buildKodyModuleBundle({
-		env: input.env,
-		baseUrl: input.baseUrl,
-		userId: input.userId,
-		sourceFiles: packageSource.files,
-		entryPoint: resolution.entryPoint,
-		rootPackageId: input.savedPackage.id,
-	})
-	await persistPublishedBundleArtifact({
-		env: input.env,
-		userId: input.userId,
-		source: packageSource.source,
-		kind: 'module',
-		artifactName: resolution.artifactName,
-		entryPoint: resolution.entryPoint,
-		mainModule: bundle.mainModule,
-		modules: bundle.modules,
-		dependencies: bundle.dependencies,
-		dynamicDependencies: bundle.dynamicDependencies,
-		packageContext: {
-			packageId: input.savedPackage.id,
-			kodyId: input.savedPackage.kodyId,
-			sourceId: input.savedPackage.sourceId,
-		},
-	})
-	const rebuilt = await loadPublishedBundleArtifactByIdentity({
-		env: input.env,
-		userId: input.userId,
-		sourceId: input.savedPackage.sourceId,
-		kind: 'module',
-		artifactName: resolution.artifactName,
-		entryPoint: resolution.entryPoint,
-	})
-	if (!rebuilt?.artifact) {
-		const moduleLabel =
-			input.selector.kind === 'export'
-				? `export "${input.selector.exportName}"`
-				: `subscription "${input.selector.topic}"`
-		throw new Error(
-			`Published bundle artifact for ${moduleLabel} could not be loaded after rebuild.`,
-		)
-	}
-	return {
-		artifact: rebuilt.artifact,
-		source: packageSource.source,
-		entryPoint: resolution.entryPoint,
-	}
-}
-
-function buildExecutionSuccessResponse(input: {
-	savedPackage: SavedPackageRecord
-	invocationName: string
-	idempotencyKey: string
-	source: string | null
-	topic: string | null
-	result: unknown
-	logs: Array<string>
-	rawContent: Array<ContentBlock> | null
-	rawContentOmitted: null | {
-		note: string
-		returnedBytes: number
-		blockCount: number
-	}
-}): PackageInvocationStoredResponse {
-	return {
-		status: 200,
-		body: {
-			ok: true,
-			package: {
-				id: input.savedPackage.id,
-				kodyId: input.savedPackage.kodyId,
-			},
-			exportName: input.invocationName,
-			source: input.source,
-			topic: input.topic,
-			idempotency: {
-				key: input.idempotencyKey,
-				replayed: false,
-			},
-			result: toJsonSafeValue(input.result),
-			logs: input.logs,
-			...(input.rawContent
-				? { rawContent: toJsonSafeValue(input.rawContent) }
-				: {}),
-			...(input.rawContentOmitted
-				? { rawContentOmitted: input.rawContentOmitted }
-				: {}),
-		},
-	}
-}
-
-function buildExecutionErrorResponse(input: {
-	savedPackage: SavedPackageRecord
-	invocationName: string
-	idempotencyKey: string
-	source: string | null
-	topic: string | null
-	error: unknown
-	logs: Array<string>
-}): PackageInvocationStoredResponse {
-	const message = getErrorMessage(input.error)
-	return {
-		status: 500,
-		body: {
-			ok: false,
-			package: {
-				id: input.savedPackage.id,
-				kodyId: input.savedPackage.kodyId,
-			},
-			exportName: input.invocationName,
-			source: input.source,
-			topic: input.topic,
-			idempotency: {
-				key: input.idempotencyKey,
-				replayed: false,
-			},
-			error: {
-				code: 'execution_failed',
-				message,
-				details: toJsonSafeValue(getExecutionErrorDetails(input.error)),
-			},
-			logs: input.logs,
-		},
-	}
-}
-
-function buildJsonErrorResponse(input: {
-	status: number
-	code: string
-	message: string
-	idempotencyKey?: string
-	replayed?: boolean
-}) {
-	return {
-		status: input.status,
-		body: {
-			ok: false,
-			error: {
-				code: input.code,
-				message: input.message,
-			},
-			...(input.idempotencyKey
-				? {
-						idempotency: {
-							key: input.idempotencyKey,
-							replayed: input.replayed ?? false,
-						},
-					}
-				: {}),
-		},
-	} satisfies PackageInvocationStoredResponse
-}
-
-function buildIdempotencyResponseUnavailable(input: {
-	idempotencyKey: string
-}) {
-	return buildJsonErrorResponse({
-		status: 409,
-		code: 'idempotency_response_unavailable',
-		message:
-			'This idempotency key already has a terminal invocation record, but its stored response could not be replayed.',
-		idempotencyKey: input.idempotencyKey,
-	})
-}
-
-function resolveExistingInvocation(input: {
-	record: NonNullable<Awaited<ReturnType<typeof getPackageInvocationByKey>>>
-	requestHash: string
-	idempotencyKey: string
-}): PackageInvocationStoredResponse {
-	if (input.record.request_hash !== input.requestHash) {
-		return buildJsonErrorResponse({
-			status: 409,
-			code: 'idempotency_mismatch',
-			message:
-				'This idempotency key has already been used for a different package invocation request.',
-			idempotencyKey: input.idempotencyKey,
-		})
-	}
-	if (input.record.status === 'in_progress') {
-		return buildJsonErrorResponse({
-			status: 409,
-			code: 'invocation_in_progress',
-			message:
-				'This idempotency key is already processing for the requested package export.',
-			idempotencyKey: input.idempotencyKey,
-		})
-	}
-	if (input.record.storedResponse) {
-		return markStoredResponseAsReplayed(input.record.storedResponse)
-	}
-	return buildIdempotencyResponseUnavailable({
-		idempotencyKey: input.idempotencyKey,
-	})
-}
-
-function resolvePackageModuleResolution(input: {
-	manifest: Awaited<ReturnType<typeof loadPackageSourceBySourceId>>['manifest']
-	selector: PackageModuleSelector
-}): PackageModuleResolution {
-	switch (input.selector.kind) {
-		case 'export': {
-			const exportName = normalizeExportName(input.selector.exportName)
-			return {
-				artifactName: exportName,
-				entryPoint: resolvePackageExportPath({
-					manifest: input.manifest,
-					exportName,
-				}),
-			}
-		}
-		case 'subscription': {
-			const topic = normalizePackageSubscriptionTopic(input.selector.topic)
-			const handler = input.manifest.kody.subscriptions?.[topic]?.handler
-			if (!handler) {
-				throw new Error(
-					`Package "${input.manifest.kody.id}" does not define subscription "${topic}".`,
-				)
-			}
-			return {
-				artifactName: buildPackageSubscriptionArtifactName(topic),
-				entryPoint: normalizePackageWorkspacePath(handler),
-			}
-		}
-		default: {
-			const selector = input.selector
-			void selector
-			throw new Error('Unhandled package module selector.')
-		}
-	}
-}
-
-function isMissingPackageModuleError(error: unknown) {
-	return (
-		error instanceof Error &&
-		(error.message.includes('does not define export') ||
-			error.message.includes('does not define a runtime target') ||
-			error.message.includes('does not define subscription'))
-	)
-}
-
-function resolveInvocationRuntimeSurface(input: {
-	selector: PackageModuleSelector
-	source: string | null
-}): PackageRuntimeSurface {
-	if (input.source === packageWorkflowInvocationSource) return 'workflow'
-	switch (input.selector.kind) {
-		case 'export':
-			return 'export'
-		case 'subscription':
-			return 'subscription'
-		default: {
-			const selector: never = input.selector
-			void selector
-			throw new Error('Unhandled package module selector.')
-		}
-	}
-}
-
-function resolveInvocationRuntimeName(input: {
-	surface: PackageRuntimeSurface
-	invocationName: string
-	topic: string | null
-}) {
-	switch (input.surface) {
-		case 'workflow':
-		case 'subscription':
-			return input.topic ?? input.invocationName
-		case 'export':
-			return input.invocationName
-		case 'app_fetch':
-		case 'app_realtime':
-		case 'service':
-		case 'job':
-		case 'retriever':
-			return input.invocationName
-		default: {
-			const surface: never = input.surface
-			void surface
-			throw new Error('Unhandled package runtime surface.')
-		}
-	}
-}
-
-async function invokeSavedPackageModule(input: {
-	env: Env
-	baseUrl: string
-	actor: PackageInvocationActor
-	savedPackage: SavedPackageRecord
-	invocationName: string
-	moduleSelector: PackageModuleSelector
-	params?: Record<string, unknown>
-	idempotencyKey: string
-	source: string | null
-	topic: string | null
-	notFoundCode: 'export_not_found' | 'subscription_not_found'
-	runtimeInvokeDepth?: number
-}) {
-	const requestHash = await createRequestHash({
-		packageId: input.savedPackage.id,
-		exportName: input.invocationName,
-		params: input.params,
-		source: input.source,
-		topic: input.topic,
-	})
-	let existing: Awaited<ReturnType<typeof getPackageInvocationByKey>>
-	try {
-		existing = await getPackageInvocationByKey({
-			db: input.env.APP_DB,
-			userId: input.actor.userId,
-			tokenId: input.actor.tokenId,
-			packageId: input.savedPackage.id,
-			exportName: input.invocationName,
-			idempotencyKey: input.idempotencyKey,
-		})
-	} catch (error) {
-		console.error('package invocation idempotency lookup failed', error)
-		return buildJsonErrorResponse({
-			status: 500,
-			code: 'idempotency_lookup_failed',
-			message:
-				'Unable to look up the package invocation idempotency record. Please retry.',
-			idempotencyKey: input.idempotencyKey,
-		})
-	}
-	if (existing) {
-		return resolveExistingInvocation({
-			record: existing,
-			requestHash,
-			idempotencyKey: input.idempotencyKey,
-		})
-	}
-
-	const invocationId = crypto.randomUUID()
-	let inserted: boolean
-	try {
-		inserted = await insertPackageInvocationRow({
-			db: input.env.APP_DB,
-			row: {
-				id: invocationId,
-				userId: input.actor.userId,
-				tokenId: input.actor.tokenId,
-				packageId: input.savedPackage.id,
-				packageKodyId: input.savedPackage.kodyId,
-				exportName: input.invocationName,
-				idempotencyKey: input.idempotencyKey,
-				requestHash,
-				source: input.source,
-				topic: input.topic,
-				status: 'in_progress',
-			},
-		})
-	} catch (error) {
-		console.error('package invocation idempotency persistence failed', error)
-		return buildJsonErrorResponse({
-			status: 500,
-			code: 'idempotency_persistence_failed',
-			message:
-				'Unable to persist the package invocation idempotency record. Please retry.',
-			idempotencyKey: input.idempotencyKey,
-		})
-	}
-	if (!inserted) {
-		let current: Awaited<ReturnType<typeof getPackageInvocationByKey>>
-		try {
-			current = await getPackageInvocationByKey({
-				db: input.env.APP_DB,
-				userId: input.actor.userId,
-				tokenId: input.actor.tokenId,
-				packageId: input.savedPackage.id,
-				exportName: input.invocationName,
-				idempotencyKey: input.idempotencyKey,
-			})
-		} catch (error) {
-			console.error('package invocation idempotency lookup failed', error)
-			return buildJsonErrorResponse({
-				status: 500,
-				code: 'idempotency_lookup_failed',
-				message:
-					'Unable to look up the package invocation idempotency record. Please retry.',
-				idempotencyKey: input.idempotencyKey,
-			})
-		}
-		if (!current) {
-			return buildJsonErrorResponse({
-				status: 500,
-				code: 'idempotency_conflict_unresolved',
-				message:
-					'Package invocation idempotency insert conflicted but no existing row was found.',
-				idempotencyKey: input.idempotencyKey,
-			})
-		}
-		return resolveExistingInvocation({
-			record: current,
-			requestHash,
-			idempotencyKey: input.idempotencyKey,
-		})
-	}
-
-	try {
-		const { artifact, source: sourceRow } = await ensureModuleArtifact({
-			env: input.env,
-			baseUrl: input.baseUrl,
-			savedPackage: input.savedPackage,
-			selector: input.moduleSelector,
-			userId: input.actor.userId,
-		})
-		const repoSource =
-			artifact.packageContext?.sourceId != null
-				? await getEntitySourceById(
-						input.env.APP_DB,
-						artifact.packageContext.sourceId,
-					)
-				: sourceRow
-		const callerContext = createMcpCallerContext({
-			baseUrl: input.baseUrl,
-			executionOrigin: 'background',
-			user: {
-				userId: input.actor.userId,
-				email: input.actor.email,
-				username: undefined,
-				displayName: input.actor.displayName,
-			},
-			storageContext: {
-				sessionId: null,
-				appId: input.savedPackage.id,
-				packageId: input.savedPackage.id,
-				storageId: buildPackageInvocationStorageId(input.savedPackage.id),
-			},
-			remoteConnectors: input.actor.remoteConnectors ?? null,
-			repoContext: repoSource ? createRepoContext(repoSource) : null,
-		})
-		const runtimeSurface = resolveInvocationRuntimeSurface({
-			selector: input.moduleSelector,
-			source: input.source,
-		})
-		const packageContext = artifact.packageContext ?? {
-			packageId: input.savedPackage.id,
-			kodyId: input.savedPackage.kodyId,
-			sourceId: input.savedPackage.sourceId,
-		}
-		const runtimeDebug: PackageRuntimeDebugContext = {
-			packageId: input.savedPackage.id,
-			kodyId: input.savedPackage.kodyId,
-			sourceId: input.savedPackage.sourceId,
-			publishedCommit: repoSource?.published_commit ?? null,
-			surface: runtimeSurface,
-			name: resolveInvocationRuntimeName({
-				surface: runtimeSurface,
-				invocationName: input.invocationName,
-				topic: input.topic,
-			}),
-			invocationId,
-			idempotencyKey: input.idempotencyKey,
-			metadata: {
-				exportName: input.invocationName,
-				source: input.source,
-				topic: input.topic,
-			},
-		}
-		const executionResult = await runBundledModuleWithRegistry(
-			input.env,
-			callerContext,
-			{
-				mainModule: artifact.mainModule,
-				modules: artifact.modules,
-				dependencies: artifact.dependencies,
-			},
-			input.params,
-			{
-				// No ambient `storage` binding: package code reaches its bucket via
-				// `packageStorage()` (granted through packageContext below). Legacy
-				// ambient use gets the structured runtime_helper_unbound hint.
-				runtimeDebug,
-				emailTools: {
-					getMessage: async (messageId) => {
-						const loaded = await getEmailMessageWithAttachmentsById({
-							db: input.env.APP_DB,
-							userId: input.actor.userId,
-							messageId,
-						})
-						if (!loaded) {
-							throw new Error(`Email message not found: ${messageId}`)
-						}
-						return {
-							id: loaded.message.id,
-							direction: loaded.message.direction,
-							inbox_id: loaded.message.inboxId,
-							thread_id: loaded.message.threadId,
-							from_address: loaded.message.fromAddress,
-							envelope_from: loaded.message.envelopeFrom,
-							to_addresses: loaded.message.toAddresses,
-							cc_addresses: loaded.message.ccAddresses,
-							bcc_addresses: loaded.message.bccAddresses,
-							reply_to_addresses: loaded.message.replyToAddresses,
-							subject: loaded.message.subject,
-							message_id_header: loaded.message.messageIdHeader,
-							in_reply_to_header: loaded.message.inReplyToHeader,
-							references: loaded.message.references,
-							headers: loaded.message.headers,
-							auth_results: loaded.message.authResults,
-							text_body: loaded.message.textBody,
-							html_body: loaded.message.htmlBody,
-							raw_size: loaded.message.rawSize,
-							processing_status: loaded.message.processingStatus,
-							provider_message_id: loaded.message.providerMessageId,
-							delivery_status: loaded.message.deliveryStatus,
-							delivery_status_at: loaded.message.deliveryStatusAt,
-							error: loaded.message.error,
-							received_at: loaded.message.receivedAt,
-							sent_at: loaded.message.sentAt,
-							created_at: loaded.message.createdAt,
-							updated_at: loaded.message.updatedAt,
-							attachments: loaded.attachments.map((attachment) => ({
-								id: attachment.id,
-								filename: attachment.filename,
-								content_type: attachment.contentType,
-								content_id: attachment.contentId,
-								disposition: attachment.disposition,
-								size: attachment.size,
-								storage_kind: attachment.storageKind,
-								storage_key: attachment.storageKey,
-								created_at: attachment.createdAt,
-							})),
-						}
-					},
-					getAttachment: async (attachmentId) => {
-						const attachment = await getEmailAttachmentById({
-							db: input.env.APP_DB,
-							blobs: input.env.EMAIL_BLOBS,
-							userId: input.actor.userId,
-							attachmentId,
-						})
-						if (!attachment) {
-							throw new Error(`Email attachment not found: ${attachmentId}`)
-						}
-						const message = await getEmailMessageById({
-							db: input.env.APP_DB,
-							userId: input.actor.userId,
-							messageId: attachment.messageId,
-						})
-						if (!message) {
-							throw new Error(
-								`Email message not found for attachment: ${attachment.messageId}`,
-							)
-						}
-						return {
-							id: attachment.id,
-							message_id: attachment.messageId,
-							filename: attachment.filename,
-							content_type: attachment.contentType,
-							content_id: attachment.contentId,
-							disposition: attachment.disposition,
-							size: attachment.size,
-							storage_kind: attachment.storageKind,
-							storage_key: attachment.storageKey,
-							created_at: attachment.createdAt,
-							message: {
-								id: message.id,
-								message_id_header: message.messageIdHeader,
-								subject: message.subject,
-							},
-							content: attachment.content,
-							content_base64: attachment.contentBase64,
-						}
-					},
-				},
-				packageContext,
-				packageInvokeTools: createPackageRuntimeInvokeTools({
-					env: input.env,
-					baseUrl: input.baseUrl,
-					callerContext,
-					packageContext,
-					parentRuntimeDebug: runtimeDebug,
-					packageInvokeDepth: input.runtimeInvokeDepth ?? 0,
-				}),
-				packageEventTools: createPackageEventTools({
-					env: input.env,
-					baseUrl: input.baseUrl,
-					callerContext,
-					packageContext,
-					parentRuntimeDebug: runtimeDebug,
-					packageInvokeDepth: input.runtimeInvokeDepth ?? 0,
-				}),
-			},
-		)
-		let response: PackageInvocationStoredResponse
-		if (executionResult.error) {
-			response = buildExecutionErrorResponse({
-				savedPackage: input.savedPackage,
-				invocationName: input.invocationName,
-				idempotencyKey: input.idempotencyKey,
-				source: input.source,
-				topic: input.topic,
-				error: executionResult.error,
-				logs: executionResult.logs ?? [],
-			})
-		} else {
-			const persistedArtifacts = persistableExecutionArtifacts(
-				executionResult.result,
-			)
-			response = buildExecutionSuccessResponse({
-				savedPackage: input.savedPackage,
-				invocationName: input.invocationName,
-				idempotencyKey: input.idempotencyKey,
-				source: input.source,
-				topic: input.topic,
-				result: persistedArtifacts.result,
-				rawContent: persistedArtifacts.rawContent,
-				rawContentOmitted: persistedArtifacts.rawContentOmitted,
-				logs: executionResult.logs ?? [],
-			})
-		}
-		await updatePackageInvocationResult({
-			db: input.env.APP_DB,
-			id: invocationId,
-			userId: input.actor.userId,
-			status: executionResult.error ? 'failed' : 'completed',
-			response,
-		})
-		return response
-	} catch (error) {
-		if (isMissingPackageModuleError(error)) {
-			const response = buildJsonErrorResponse({
-				status: 404,
-				code: input.notFoundCode,
-				message: getErrorMessage(error),
-				idempotencyKey: input.idempotencyKey,
-			})
-			await updatePackageInvocationResult({
-				db: input.env.APP_DB,
-				id: invocationId,
-				userId: input.actor.userId,
-				status: 'failed',
-				response,
-			}).catch(() => {
-				// Best effort; preserve the original invocation error.
-			})
-			return response
-		}
-		const response = buildJsonErrorResponse({
-			status: 500,
-			code: 'invocation_failed',
-			message: getErrorMessage(error),
-			idempotencyKey: input.idempotencyKey,
-		})
-		await updatePackageInvocationResult({
-			db: input.env.APP_DB,
-			id: invocationId,
-			userId: input.actor.userId,
-			status: 'failed',
-			response,
-		}).catch(() => {
-			// Best effort; preserve the original invocation error.
-		})
-		return response
-	}
+	},
 }
 
 export function createPackageRuntimeInvokeTools(input: {
@@ -1304,108 +55,10 @@ export function createPackageRuntimeInvokeTools(input: {
 	parentRuntimeDebug?: PackageRuntimeDebugContext | null
 	packageInvokeDepth?: number
 }): PackageInvokeTools {
-	return createPackageInvokeTools({
+	return createPackageRuntimeInvokeToolsWithToolFactories({
 		...input,
-		callerKind: 'package',
+		toolFactories: packageRuntimeToolFactories,
 	})
-}
-
-function parsePackageEventDispatchInput(rawInput: PackageEventDispatchInput) {
-	const input =
-		rawInput && typeof rawInput === 'object' && !Array.isArray(rawInput)
-			? rawInput
-			: {}
-	const topic =
-		typeof input.topic === 'string'
-			? normalizePackageSubscriptionTopic(input.topic)
-			: ''
-	if (!topic) {
-		throw new Error('events.dispatch requires a non-empty topic.')
-	}
-	const idempotencyKey =
-		typeof input.idempotencyKey === 'string' ? input.idempotencyKey.trim() : ''
-	if (!idempotencyKey) {
-		throw new Error('events.dispatch requires a non-empty idempotencyKey.')
-	}
-	const payload = input.payload ?? {}
-	if (!payload || typeof payload !== 'object' || Array.isArray(payload)) {
-		throw new Error(
-			'events.dispatch payload must be a JSON object when provided.',
-		)
-	}
-	return {
-		topic,
-		idempotencyKey,
-		payload: payload as Record<string, unknown>,
-	}
-}
-
-async function buildPackageEventSubscriptionIdempotencyKey(input: {
-	sourcePackageId: string
-	subscriberPackageId: string
-	topic: string
-	idempotencyKey: string
-}) {
-	const digest = await crypto.subtle.digest(
-		'SHA-256',
-		new TextEncoder().encode(canonicalJsonStringify(input)),
-	)
-	return [
-		'pkgevent',
-		input.sourcePackageId,
-		input.subscriberPackageId,
-		input.topic,
-		toHex(new Uint8Array(digest)).slice(0, 24),
-	].join(':')
-}
-
-function readInvocationError(response: PackageInvocationResponse) {
-	const errorRecord =
-		(response.body['error'] as Record<string, unknown> | undefined) ?? {}
-	return {
-		code: String(errorRecord['code'] ?? 'package_subscription_failed'),
-		message: String(
-			errorRecord['message'] ??
-				`Package subscription failed with HTTP ${response.status}.`,
-		),
-	}
-}
-
-async function loadMatchingPackageEventSubscriptions(input: {
-	env: Env
-	baseUrl: string
-	userId: string
-	topic: string
-}) {
-	const savedPackages = await listSavedPackagesByUserId(input.env.APP_DB, {
-		userId: input.userId,
-	})
-	const settled = await Promise.all(
-		savedPackages.map(async (savedPackage) => {
-			const loaded = await loadPackageManifestBySourceId({
-				env: input.env,
-				baseUrl: input.baseUrl,
-				userId: input.userId,
-				sourceId: savedPackage.sourceId,
-			}).catch((error) => {
-				throw new Error(
-					`Failed to load package manifest for package event dispatch: ${savedPackage.kodyId} (${savedPackage.id}).`,
-					{ cause: error },
-				)
-			})
-			const subscription = listPackageSubscriptions(loaded.manifest).find(
-				(candidate) => candidate.topic === input.topic,
-			)
-			if (!subscription) return null
-			return {
-				savedPackage,
-				subscription,
-			}
-		}),
-	)
-	return settled.filter(
-		(entry): entry is LoadedPackageEventSubscription => entry !== null,
-	)
 }
 
 export function createPackageEventTools(input: {
@@ -1416,112 +69,10 @@ export function createPackageEventTools(input: {
 	parentRuntimeDebug?: PackageRuntimeDebugContext | null
 	packageInvokeDepth?: number
 }): PackageEventTools {
-	return {
-		dispatch: async (rawInput) => {
-			const user = input.callerContext.user
-			if (!user?.userId) {
-				throw new Error('events.dispatch requires an authenticated user.')
-			}
-			if (!input.packageContext) {
-				throw new Error('events.dispatch requires a package runtime context.')
-			}
-			const packageContext = input.packageContext
-			const packageInvokeDepth = input.packageInvokeDepth ?? 0
-			if (packageInvokeDepth >= maxPackageRuntimeInvokeDepth) {
-				throw new Error(
-					`events.dispatch exceeded the maximum nested invocation depth (${maxPackageRuntimeInvokeDepth}).`,
-				)
-			}
-			const request = parsePackageEventDispatchInput(rawInput)
-			const sourceId = packageContext.sourceId?.trim()
-			if (!sourceId) {
-				throw new Error('events.dispatch requires a published package source.')
-			}
-			const sourceManifest = await loadPackageManifestBySourceId({
-				env: input.env,
-				baseUrl: input.baseUrl,
-				userId: user.userId,
-				sourceId,
-			})
-			const declaredEvent = listPackageEmittedEvents(
-				sourceManifest.manifest,
-			).find((event) => event.topic === request.topic)
-			if (!declaredEvent) {
-				throw new Error(
-					`Package "${packageContext.kodyId}" does not declare emitted event "${request.topic}" in package.json#kody.emits.`,
-				)
-			}
-			const subscriptions = await loadMatchingPackageEventSubscriptions({
-				env: input.env,
-				baseUrl: input.baseUrl,
-				userId: user.userId,
-				topic: request.topic,
-			})
-			const envelope = {
-				event: request.topic,
-				source: {
-					type: 'package',
-					package_id: packageContext.packageId,
-					kody_id: packageContext.kodyId,
-				},
-				idempotency_key: request.idempotencyKey,
-				payload: request.payload,
-			}
-			const subscribers = []
-			for (const { savedPackage, subscription } of subscriptions) {
-				const response = await invokePackageSubscription({
-					env: input.env,
-					baseUrl: input.baseUrl,
-					savedPackage,
-					topic: request.topic,
-					params: envelope,
-					idempotencyKey: await buildPackageEventSubscriptionIdempotencyKey({
-						sourcePackageId: packageContext.packageId,
-						subscriberPackageId: savedPackage.id,
-						topic: request.topic,
-						idempotencyKey: request.idempotencyKey,
-					}),
-					source: `package:${packageContext.kodyId}`,
-					actorTokenId: `${internalPackageEventSubscriptionTokenId}:${packageContext.packageId}`,
-					actorDisplayName: `package:${packageContext.kodyId}`,
-					runtimeInvokeDepth: packageInvokeDepth + 1,
-				})
-				const replayed =
-					(response.body['idempotency'] as { replayed?: unknown } | undefined)
-						?.replayed === true
-				const status =
-					response.status >= 200 && response.status < 400
-						? replayed
-							? 'replayed'
-							: 'completed'
-						: 'failed'
-				subscribers.push({
-					packageId: savedPackage.id,
-					kodyId: savedPackage.kodyId,
-					handler: subscription.handler,
-					status,
-					...(status === 'failed'
-						? { error: readInvocationError(response) }
-						: {}),
-				})
-			}
-			const failed = subscribers.filter(
-				(subscriber) => subscriber.status === 'failed',
-			).length
-			return {
-				topic: request.topic,
-				source: {
-					type: 'package',
-					packageId: packageContext.packageId,
-					kodyId: packageContext.kodyId,
-				},
-				idempotencyKey: request.idempotencyKey,
-				subscribers,
-				delivered: subscribers.length - failed,
-				failed,
-			}
-		},
-	}
+	return createPackageEventToolsWithToolFactories({
+		...input,
+		toolFactories: packageRuntimeToolFactories,
+	})
 }
 
 export function createExecutePackageInvokeTools(input: {
@@ -1533,294 +84,9 @@ export function createExecutePackageInvokeTools(input: {
 	/** MCP execute conversation id for agent-facing popularity recording. */
 	conversationId?: string | null
 }): PackageInvokeTools {
-	return createPackageInvokeTools({
+	return createExecutePackageInvokeToolsWithToolFactories({
 		...input,
-		packageContext: null,
-		callerKind: 'execute',
-		conversationId: input.conversationId ?? null,
-	})
-}
-
-function createPackageInvokeTools(input: {
-	env: Env
-	baseUrl: string
-	callerContext: ReturnType<typeof createMcpCallerContext>
-	packageContext: PackageRuntimeContext | null
-	parentRuntimeDebug?: PackageRuntimeDebugContext | null
-	packageInvokeDepth?: number
-	callerKind: 'package' | 'execute'
-	conversationId?: string | null
-}): PackageInvokeTools {
-	let autoIdempotencySequence = 0
-	const requireRuntimeCaller = (operationName: string) => {
-		const user = input.callerContext.user
-		if (!user?.userId) {
-			throw new Error(`${operationName} requires an authenticated user.`)
-		}
-		if (input.callerKind === 'package' && !input.packageContext) {
-			throw new Error(`${operationName} requires a package runtime context.`)
-		}
-		return { user, packageContext: input.packageContext }
-	}
-	const invoke = async (rawInput: PackageInvokeInput) => {
-		const { user, packageContext } = requireRuntimeCaller('packages.invoke')
-		const packageInvokeDepth = input.packageInvokeDepth ?? 0
-		if (packageInvokeDepth >= maxPackageRuntimeInvokeDepth) {
-			throw new Error(
-				`packages.invoke exceeded the maximum nested invocation depth (${maxPackageRuntimeInvokeDepth}).`,
-			)
-		}
-		const request = parsePackageInvokeInput(rawInput)
-		autoIdempotencySequence += 1
-		const idempotencyKey =
-			request.idempotencyKey ??
-			(await createAutoPackageInvokeIdempotencyKey({
-				callerPackageContext: packageContext,
-				parentRuntimeDebug: input.parentRuntimeDebug ?? null,
-				sequence: autoIdempotencySequence,
-				request,
-			}))
-		const response = packageContext
-			? await invokePackageExportForPackageRuntime({
-					env: input.env,
-					baseUrl: input.baseUrl,
-					caller: {
-						userId: user.userId,
-						email: user.email ?? '',
-						displayName: user.displayName ?? '',
-						remoteConnectors: input.callerContext.remoteConnectors ?? null,
-						packageContext,
-					},
-					request: {
-						packageIdOrKodyId: request.packageIdOrKodyId,
-						exportName: request.exportName,
-						params: request.params,
-						idempotencyKey,
-						source: `package:${packageContext.kodyId}`,
-						topic: request.topic,
-					},
-					runtimeInvokeDepth: packageInvokeDepth + 1,
-				})
-			: await invokePackageExportForExecuteRuntime({
-					env: input.env,
-					baseUrl: input.baseUrl,
-					caller: {
-						userId: user.userId,
-						email: user.email ?? '',
-						displayName: user.displayName ?? '',
-						remoteConnectors: input.callerContext.remoteConnectors ?? null,
-					},
-					request: {
-						packageIdOrKodyId: request.packageIdOrKodyId,
-						exportName: request.exportName,
-						params: request.params,
-						idempotencyKey,
-						topic: request.topic,
-					},
-					runtimeInvokeDepth: packageInvokeDepth + 1,
-					conversationId: input.conversationId ?? null,
-				})
-		if (response.status >= 200 && response.status < 400) {
-			return response.body['result']
-		}
-		const errorRecord =
-			(response.body['error'] as Record<string, unknown> | undefined) ?? {}
-		const code = String(errorRecord['code'] ?? 'package_invocation_failed')
-		const message = String(
-			errorRecord['message'] ??
-				`Package invocation failed with HTTP ${response.status}.`,
-		)
-		const error = new Error(`[${code}] ${message}`) as Error & {
-			code?: string
-			status?: number
-			response?: PackageInvocationResponse
-		}
-		error.code = code
-		error.status = response.status
-		error.response = response
-		throw error
-	}
-	return {
-		check: async (rawInput) => {
-			const { user } = requireRuntimeCaller('packages.check')
-			return await checkPackageInvokeForRuntime({
-				env: input.env,
-				baseUrl: input.baseUrl,
-				operationName: 'packages.check',
-				userId: user.userId,
-				rawInput,
-			})
-		},
-		invoke,
-		invokeChecked: async (rawInput) => {
-			const { user } = requireRuntimeCaller('packages.invokeChecked')
-			const check = await checkPackageInvokeForRuntime({
-				env: input.env,
-				baseUrl: input.baseUrl,
-				operationName: 'packages.invokeChecked',
-				userId: user.userId,
-				rawInput,
-			})
-			if (!check.ok) {
-				const error = new Error(
-					`packages.invokeChecked check failed: ${check.message}`,
-				) as Error & {
-					check?: PackageInvokeCheckResult
-				}
-				error.check = check
-				throw error
-			}
-			return await invoke(check.invoke)
-		},
-	}
-}
-
-async function invokePackageExportForExecuteRuntime(input: {
-	env: Env
-	baseUrl: string
-	caller: {
-		userId: string
-		email: string
-		displayName: string
-		remoteConnectors?: Array<RemoteConnectorRef> | null
-	}
-	request: PackageInvocationRequest
-	runtimeInvokeDepth?: number
-	conversationId?: string | null
-}): Promise<PackageInvocationResponse> {
-	const packageIdOrKodyId = input.request.packageIdOrKodyId.trim()
-	if (!packageIdOrKodyId) {
-		return buildJsonErrorResponse({
-			status: 400,
-			code: 'invalid_package',
-			message: 'Package id or kody id is required.',
-		})
-	}
-	const exportName = normalizeExportName(input.request.exportName)
-	const idempotencyKey = input.request.idempotencyKey.trim()
-	if (!idempotencyKey) {
-		return buildJsonErrorResponse({
-			status: 400,
-			code: 'missing_idempotency_key',
-			message: 'Package invocations require a non-empty idempotencyKey.',
-		})
-	}
-	const savedPackage = await resolveSavedPackage({
-		db: input.env.APP_DB,
-		userId: input.caller.userId,
-		packageIdOrKodyId,
-	})
-	if (!savedPackage) {
-		return buildJsonErrorResponse({
-			status: 404,
-			code: 'package_not_found',
-			message: `Saved package "${packageIdOrKodyId}" was not found for this user.`,
-			idempotencyKey,
-		})
-	}
-	const conversationId = input.conversationId?.trim()
-	if (conversationId) {
-		// Best-effort; must not affect invoke. Conversation-unique upsert.
-		void recordAgentPackageConversationUse(input.env, {
-			userId: input.caller.userId,
-			packageId: savedPackage.id,
-			conversationId,
-		})
-	}
-	return await invokeSavedPackageModule({
-		env: input.env,
-		baseUrl: input.baseUrl,
-		actor: {
-			tokenId: internalExecuteRuntimeInvokeTokenId,
-			userId: input.caller.userId,
-			email: input.caller.email,
-			displayName: input.caller.displayName || input.caller.email || 'execute',
-			remoteConnectors: input.caller.remoteConnectors ?? null,
-		},
-		savedPackage,
-		invocationName: exportName,
-		moduleSelector: {
-			kind: 'export',
-			exportName,
-		},
-		params: input.request.params,
-		idempotencyKey,
-		source: 'execute',
-		topic: normalizeNullableString(input.request.topic),
-		notFoundCode: 'export_not_found',
-		runtimeInvokeDepth: input.runtimeInvokeDepth ?? 0,
-	})
-}
-
-async function invokePackageExportForPackageRuntime(input: {
-	env: Env
-	baseUrl: string
-	caller: {
-		userId: string
-		email: string
-		displayName: string
-		remoteConnectors?: Array<RemoteConnectorRef> | null
-		packageContext: PackageRuntimeContext
-	}
-	request: PackageInvocationRequest
-	runtimeInvokeDepth?: number
-}): Promise<PackageInvocationResponse> {
-	const packageIdOrKodyId = input.request.packageIdOrKodyId.trim()
-	if (!packageIdOrKodyId) {
-		return buildJsonErrorResponse({
-			status: 400,
-			code: 'invalid_package',
-			message: 'Package id or kody id is required.',
-		})
-	}
-	const exportName = normalizeExportName(input.request.exportName)
-	const idempotencyKey = input.request.idempotencyKey.trim()
-	if (!idempotencyKey) {
-		return buildJsonErrorResponse({
-			status: 400,
-			code: 'missing_idempotency_key',
-			message: 'Package invocations require a non-empty idempotencyKey.',
-		})
-	}
-	const savedPackage = await resolveSavedPackage({
-		db: input.env.APP_DB,
-		userId: input.caller.userId,
-		packageIdOrKodyId,
-	})
-	if (!savedPackage) {
-		return buildJsonErrorResponse({
-			status: 404,
-			code: 'package_not_found',
-			message: `Saved package "${packageIdOrKodyId}" was not found for this user.`,
-			idempotencyKey,
-		})
-	}
-	return await invokeSavedPackageModule({
-		env: input.env,
-		baseUrl: input.baseUrl,
-		actor: {
-			tokenId: `${internalPackageRuntimeInvokeTokenId}:${input.caller.packageContext.packageId}`,
-			userId: input.caller.userId,
-			email: input.caller.email,
-			displayName:
-				input.caller.displayName ||
-				`package:${input.caller.packageContext.kodyId}`,
-			remoteConnectors: input.caller.remoteConnectors ?? null,
-		},
-		savedPackage,
-		invocationName: exportName,
-		moduleSelector: {
-			kind: 'export',
-			exportName,
-		},
-		params: input.request.params,
-		idempotencyKey,
-		source:
-			normalizeNullableString(input.request.source) ??
-			`package:${input.caller.packageContext.kodyId}`,
-		topic: normalizeNullableString(input.request.topic),
-		notFoundCode: 'export_not_found',
-		runtimeInvokeDepth: input.runtimeInvokeDepth ?? 0,
+		toolFactories: packageRuntimeToolFactories,
 	})
 }
 
@@ -1831,85 +97,9 @@ export async function invokePackageExport(input: {
 	request: PackageInvocationRequest
 	runtimeInvokeDepth?: number
 }): Promise<PackageInvocationResponse> {
-	const packageIdOrKodyId = input.request.packageIdOrKodyId.trim()
-	if (!packageIdOrKodyId) {
-		return buildJsonErrorResponse({
-			status: 400,
-			code: 'invalid_package',
-			message: 'Package id or kody id is required.',
-		})
-	}
-	const exportName = normalizeExportName(input.request.exportName)
-	const idempotencyKey = input.request.idempotencyKey.trim()
-	if (!idempotencyKey) {
-		return buildJsonErrorResponse({
-			status: 400,
-			code: 'missing_idempotency_key',
-			message: 'Package invocations require a non-empty idempotencyKey.',
-		})
-	}
-	const source = normalizeNullableString(input.request.source)
-	const topic = normalizeNullableString(input.request.topic)
-	if (!tokenAllowsSource({ token: input.token, source })) {
-		return buildJsonErrorResponse({
-			status: 403,
-			code: 'source_not_allowed',
-			message: 'This token is not allowed to invoke the requested source.',
-			idempotencyKey,
-		})
-	}
-	const savedPackage = await resolveSavedPackage({
-		db: input.env.APP_DB,
-		userId: input.token.userId,
-		packageIdOrKodyId,
-	})
-	if (!savedPackage) {
-		return buildJsonErrorResponse({
-			status: 404,
-			code: 'package_not_found',
-			message: `Saved package "${packageIdOrKodyId}" was not found for this user.`,
-			idempotencyKey,
-		})
-	}
-	if (!tokenAllowsPackage({ token: input.token, savedPackage })) {
-		return buildJsonErrorResponse({
-			status: 403,
-			code: 'package_not_allowed',
-			message: 'This token is not allowed to invoke the requested package.',
-			idempotencyKey,
-		})
-	}
-	if (!tokenAllowsExport({ token: input.token, exportName })) {
-		return buildJsonErrorResponse({
-			status: 403,
-			code: 'export_not_allowed',
-			message: `This token is not allowed to invoke export "${exportName}".`,
-			idempotencyKey,
-		})
-	}
-
-	return await invokeSavedPackageModule({
-		env: input.env,
-		baseUrl: input.baseUrl,
-		actor: {
-			tokenId: input.token.tokenId,
-			userId: input.token.userId,
-			email: input.token.email,
-			displayName: input.token.displayName,
-			remoteConnectors: input.token.remoteConnectors ?? null,
-		},
-		savedPackage,
-		invocationName: exportName,
-		moduleSelector: {
-			kind: 'export',
-			exportName,
-		},
-		params: input.request.params,
-		idempotencyKey,
-		source,
-		topic,
-		notFoundCode: 'export_not_found',
-		runtimeInvokeDepth: input.runtimeInvokeDepth ?? 0,
+	return await invokePackageExportWithToolFactories({
+		...input,
+		toolFactories: packageRuntimeToolFactories,
 	})
 }
 
@@ -1925,37 +115,8 @@ export async function invokePackageSubscription(input: {
 	actorDisplayName?: string
 	runtimeInvokeDepth?: number
 }) {
-	const topic = normalizePackageSubscriptionTopic(input.topic)
-	const idempotencyKey = input.idempotencyKey.trim()
-	if (!idempotencyKey) {
-		return buildJsonErrorResponse({
-			status: 400,
-			code: 'missing_idempotency_key',
-			message:
-				'Package subscription invocations require a non-empty idempotencyKey.',
-		})
-	}
-	return await invokeSavedPackageModule({
-		env: input.env,
-		baseUrl: input.baseUrl,
-		actor: {
-			tokenId: input.actorTokenId ?? internalEmailSubscriptionTokenId,
-			userId: input.savedPackage.userId,
-			email: '',
-			displayName:
-				input.actorDisplayName ?? `package:${input.savedPackage.kodyId}`,
-		},
-		savedPackage: input.savedPackage,
-		invocationName: buildPackageSubscriptionArtifactName(topic),
-		moduleSelector: {
-			kind: 'subscription',
-			topic,
-		},
-		params: input.params,
-		idempotencyKey,
-		source: normalizeNullableString(input.source) ?? 'email',
-		topic,
-		notFoundCode: 'subscription_not_found',
-		runtimeInvokeDepth: input.runtimeInvokeDepth ?? 0,
+	return await invokePackageSubscriptionWithToolFactories({
+		...input,
+		toolFactories: packageRuntimeToolFactories,
 	})
 }
